Package update + handle "and"
Ranges with "and" did not work as expected.

    > 2 and < 10

People think that this is a logical AND, finding all items greater than
2 that are also less than 10. Before, this would treat it as a lexical
AND (meaning "English") and that treated the ranges as a logical OR. To
explain the difference, imagine the sentence:

    I want to talk to all children and all adults.

As a logical statement, you would like to talk to people who

    = child and = adult

In that case, a logical AND would not match anyone because a person
can't be a child and an adult at the same time.

Going forward, this will remain as a logical AND.